abr-shell-ui/src/test/kotlin/com/d20charactersheet/adventurebookresolver/shellui/command/CreateCommandTest.kt
package com.d20charactersheet.adventurebookresolver.shellui.command

import com.d20charactersheet.adventurebookresolver.shellui.shell.BookPromptProvider
import org.assertj.core.api.Assertions.assertThat
import org.junit.jupiter.api.Test
import org.junit.jupiter.api.extension.ExtendWith
import org.springframework.beans.factory.annotation.Autowired
import org.springframework.boot.test.context.SpringBootTest
import org.springframework.shell.jline.InteractiveShellApplicationRunner
import org.springframework.shell.jline.ScriptShellApplicationRunner
import org.springframework.test.context.junit.jupiter.SpringExtension

@ExtendWith(SpringExtension::class)
@SpringBootTest(properties = [
    InteractiveShellApplicationRunner.SPRING_SHELL_INTERACTIVE_ENABLED + "=false",
    ScriptShellApplicationRunner.SPRING_SHELL_SCRIPT_ENABLED + "=false"])
class CreateCommandTest {

    @Autowired
    lateinit var promptProvider: BookPromptProvider

    @Autowired
    lateinit var underTest: CreateCommand

    @Test
    internal fun `create new book`() {
        // Act
        underTest.create("my new book")

        // Assert
        assertThat(promptProvider.prompt.toString()).isEqualTo("my new book>")

    }
}